+++ AngularSPA.DataRepository/Models/OrderStatus.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -10,7 +11,7 @@ namespace AngularSPA.DataRepository.Models
     public class OrderStatus
     {
         [Key]
-        public byte OrderStatusId { get; set; }
+        public int OrderStatusId { get; set; }
         [MaxLength(150)]
         public string OrderStatusName { get; set; }
     }
